Fix compiling errors in centos 5/6/7

+++ cuda/cudata.cu
@@ -149,15 +149,15 @@ int gpu_init(const PARMS_T *parms, int *gpus, int ngpu){
 	    mem_minimum+=sizeof(Real)*nps*nxn*nyn;
 	}
 	if(parms->gpu.evl){
-	    mem_minimum+=sizeof(Real)*parms->evl.nevl*pow(parms->aper.d/parms->evl.dx, 2);
+	    mem_minimum+=sizeof(Real)*parms->evl.nevl*(long)pow(parms->aper.d/parms->evl.dx, 2);
 	}
 	if(parms->gpu.wfs){
 	    for(int ipowfs=0; ipowfs<parms->npowfs; ipowfs++){
-		mem_minimum+=sizeof(Real)*parms->powfs[ipowfs].nwfs*pow(parms->aper.d/parms->powfs[ipowfs].dx, 2)*4;
+		mem_minimum+=sizeof(Real)*parms->powfs[ipowfs].nwfs*(long)pow(parms->aper.d/parms->powfs[ipowfs].dx, 2)*4;
 	    }
 	}
 	if(parms->gpu.tomo || parms->gpu.fit){
-	    mem_minimum+=sizeof(Real)*parms->atmr.nps*pow(parms->aper.d*parms->tomo.pos/parms->atmr.dx, 2)*4;
+	    mem_minimum+=sizeof(Real)*parms->atmr.nps*(long)pow(parms->aper.d*parms->tomo.pos/parms->atmr.dx, 2)*4;
 	}
 	mem_minimum*=2;
 	if(mem_minimum==0){//gpu is disabled
